+++ apps/rush-lib/src/logic/base/BaseShrinkwrapFile.ts
@@ -3,12 +3,13 @@
 
 import * as colors from 'colors';
 import * as semver from 'semver';
-import { PackageName, FileSystem } from '@rushstack/node-core-library';
+import { FileSystem } from '@rushstack/node-core-library';
 
 import { RushConstants } from '../../logic/RushConstants';
 import { DependencySpecifier } from '../DependencySpecifier';
 import { IPolicyValidatorOptions } from '../policy/PolicyValidator';
 import { PackageManagerOptionsConfigurationBase } from '../../api/RushConfiguration';
+import { PackageNameParsers } from '../../api/PackageNameParsers';
 
 /**
  * This class is a parser for both npm's npm-shrinkwrap.json and pnpm's pnpm-lock.yaml file formats.
@@ -85,9 +86,10 @@ export abstract class BaseShrinkwrapFile {
    *
    * @virtual
    */
-  public tryEnsureCompatibleDependency(dependencySpecifier: DependencySpecifier, tempProjectName: string): boolean {
+  public tryEnsureCompatibleDependency(dependencySpecifier: DependencySpecifier,
+    tempProjectName: string, tryReusingPackageVersionsFromShrinkwrap: boolean = true): boolean {
     const shrinkwrapDependency: DependencySpecifier | undefined =
-      this.tryEnsureDependencyVersion(dependencySpecifier, tempProjectName);
+      this.tryEnsureDependencyVersion(dependencySpecifier, tempProjectName, tryReusingPackageVersionsFromShrinkwrap);
     if (!shrinkwrapDependency) {
       return false;
     }
@@ -105,7 +107,7 @@ export abstract class BaseShrinkwrapFile {
 
   /** @virtual */
   protected abstract tryEnsureDependencyVersion(dependencySpecifier: DependencySpecifier,
-    tempProjectName: string): DependencySpecifier | undefined;
+    tempProjectName: string, tryReusingPackageVersionsFromShrinkwrap: boolean): DependencySpecifier | undefined;
 
   /** @virtual */
   protected abstract getTopLevelDependencyVersion(dependencyName: string): DependencySpecifier | undefined;
@@ -113,11 +115,11 @@ export abstract class BaseShrinkwrapFile {
   /** @virtual */
   protected abstract serialize(): string;
 
-  protected _getTempProjectNames(dependencies: { [key: string]: {} } ): ReadonlyArray<string> {
+  protected _getTempProjectNames(dependencies: { [key: string]: {} }): ReadonlyArray<string> {
     const result: string[] = [];
     for (const key of Object.keys(dependencies)) {
       // If it starts with @rush-temp, then include it:
-      if (PackageName.getScope(key) === RushConstants.rushTempNpmScope) {
+      if (PackageNameParsers.permissive.getScope(key) === RushConstants.rushTempNpmScope) {
         result.push(key);
       }
     }