+++ webpack/localization-plugin/src/utilities/ResxReader.ts
@@ -111,6 +111,14 @@ export class ResxReader {
                   childNode
                 );
               } else {
+                if (locFile.hasOwnProperty(stringName)) {
+                  ResxReader._logErrorWithLocation(
+                    options,
+                    `Duplicate string value "${stringName}"`,
+                    childNode
+                  );
+                }
+
                 const locString: ILocalizedString | undefined = ResxReader._readDataElement(options, childNode);
 
                 if (locString) {
@@ -274,7 +282,7 @@ export class ResxReader {
             break;
           }
 
-          foundText = childNode.type === 'text' ? childNode.text.trim() : childNode.cdata;
+          foundText = childNode.type === 'text' ? childNode.text : childNode.cdata;
           break;
         }