+++ libraries/load-themed-styles/CHANGELOG.md
@@ -1,6 +1,41 @@
 # Change Log - @microsoft/load-themed-styles
 
-This log was last generated on Wed, 08 Apr 2020 04:07:33 GMT and should not be manually modified.
+This log was last generated on Fri, 22 May 2020 15:08:42 GMT and should not be manually modified.
+
+## 1.10.50
+Fri, 22 May 2020 15:08:42 GMT
+
+*Version update only*
+
+## 1.10.49
+Thu, 21 May 2020 23:09:44 GMT
+
+*Version update only*
+
+## 1.10.48
+Thu, 21 May 2020 15:41:59 GMT
+
+*Version update only*
+
+## 1.10.47
+Tue, 19 May 2020 15:08:19 GMT
+
+*Version update only*
+
+## 1.10.46
+Fri, 15 May 2020 08:10:59 GMT
+
+*Version update only*
+
+## 1.10.45
+Wed, 06 May 2020 08:23:45 GMT
+
+*Version update only*
+
+## 1.10.44
+Sat, 02 May 2020 00:08:16 GMT
+
+*Version update only*
 
 ## 1.10.43
 Wed, 08 Apr 2020 04:07:33 GMT