+++ stack/rush-stack-compiler-3.5/CHANGELOG.md
@@ -1,6 +1,36 @@
 # Change Log - @microsoft/rush-stack-compiler-3.5
 
-This log was last generated on Wed, 08 Apr 2020 04:07:33 GMT and should not be manually modified.
+This log was last generated on Fri, 22 May 2020 15:08:43 GMT and should not be manually modified.
+
+## 0.4.15
+Fri, 22 May 2020 15:08:43 GMT
+
+*Version update only*
+
+## 0.4.14
+Thu, 21 May 2020 23:09:44 GMT
+
+*Version update only*
+
+## 0.4.13
+Thu, 21 May 2020 15:42:00 GMT
+
+*Version update only*
+
+## 0.4.12
+Tue, 19 May 2020 15:08:20 GMT
+
+*Version update only*
+
+## 0.4.11
+Fri, 15 May 2020 08:10:59 GMT
+
+*Version update only*
+
+## 0.4.10
+Wed, 06 May 2020 08:23:45 GMT
+
+*Version update only*
 
 ## 0.4.9
 Wed, 08 Apr 2020 04:07:33 GMT